samples/index_tuning_sample/requirements.txt
langchain-google-alloydb-pg
langchain
langchain-google-vertexai
google-cloud-alloydb-connector[pg8000]